+++ tybotRoute/tiledeskChatbotPlugs/directives/DirIntent.js
@@ -2,6 +2,7 @@
 let axios = require('axios');
 let https = require("https");
 const { v4: uuidv4 } = require('uuid');
+const ms = require('minimist-string');
 
 class DirIntent {